Project config improvements + warnings & fixes
[UPDATED]   Dependency configuration.
[ADDED]     Warning flags for all possible warnings.
[FIXED]     Signedness issues.

+++ Pearl-Crypto/PearlSCrypt.m
@@ -56,7 +56,7 @@
     [super dealloc];
 }
 
-static BOOL checkResult(NSUInteger resultCode) {
+static BOOL checkResult(int resultCode) {
     
     if (resultCode == 0)
         return YES;
